pyproject.toml
[project]
name = "mindstream"
version = "0.1.0"
description = "Real-time EEG visualization with BlueMuse and Muse2"
requires-python = ">=3.14"
dependencies = [
    "pylsl>=1.16.0",
    "pygame-ce>=2.5.0",
    "numpy>=1.24.0",
]

[project.scripts]
mindstream = "mindstream.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mindstream"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.9",
    "pytest>=8.0.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "RUF003", # ambiguous characters in comments (Japanese text)
    "TC003",  # Move imports into type-checking block (false positive for runtime usage)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ty.environment]
extra-paths = ["src"]
